requirements.txt
python-telegram-bot
google-generativeai
SQLAlchemy
python-dotenv